qs - ignore spec symbols
